Build notpecl statically
So far notpecl was not completely build statically, as it was still
depending on glibc. It's now build through a Docker image based on
Alpine. As such, it's statically linked with musl and thus doesn't
depend on glibc any more.

This is needed to be able to distribute a single binary compatible with
both Debian and Alpine (which is a nice-to-have for zbuild).